+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+- Support NodeJS on ARM builds (https://github.com/heroku/heroku-buildpack-ruby/pull/1493)
 
 
 ## [v279] - 2024-09-04